# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BossaBox Challenge Documentation",
        "contact": {
            "email": "gustavohgustavo@hotmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/tags": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get all Tag",
                "description": "This endpoint return a collection of Tag.",
                "operationId": "9a1e974f6a3e825e250969bb8c1a005c",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Value 1"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": null
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-03-08T23:04:11.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Save a new Tag",
                "description": "This endpoint save a new Tag at database.",
                "operationId": "6c57d0bfce57c15ea9cd8d1e3224aacf",
                "requestBody": {
                    "content": {
                        "json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Value 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Value 1"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": null
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-03-08T23:04:11.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "O nome da categoria deve ter mais de 3 caracteres e não pode ser nulo."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Tag já cadastrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Delete Tag",
                "description": "This endpoint delete a Tag from database.",
                "operationId": "b9bc6aaf2cade633e8633d4cdc68526b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Tag não encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tools": {
            "get": {
                "tags": [
                    "Tool"
                ],
                "summary": "Get all Tools",
                "description": "This endpoint returns a collection of Tools filtered by Tag.",
                "operationId": "adf133ed57792996d46fca967a06f518",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Value 1"
                                                    },
                                                    "link": {
                                                        "type": "string",
                                                        "example": "www.value1.com"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Value 1 is a Tool very powerful"
                                                    },
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Tag"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tool"
                ],
                "summary": "Save a new Tool",
                "description": "This endpoint saves a new Tool in the database.",
                "operationId": "02add7a4c738fe36d99a1290123bcfc8",
                "requestBody": {
                    "content": {
                        "json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "link",
                                    "description",
                                    "tags"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Value 1"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "www.value1.com"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Value 1 is a powerful tool"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tag"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Value 1"
                                        },
                                        "link": {
                                            "type": "string",
                                            "example": "www.value1.com"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Value 1 is a powerful tool"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Tag1"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "O atibuto da ferramenta deve ter mais de 3 caracteres e não pode ser nulo."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "As tags a seguir não existem: value 1, value 2, por favor use as tags já cadastradas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tools/{id}": {
            "delete": {
                "tags": [
                    "Tool"
                ],
                "summary": "Delete Tool",
                "description": "This endpoint delete a Tool from the database.",
                "operationId": "b07ffff264ad7e84a12db9a55462693b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ferramenta não encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}